pyproject.toml
[project]
name = "llm-json-streaming"
version = "0.1.0"
description = "A unified interface for streaming structured JSON from OpenAI and Anthropic."
readme = "README.md"
requires-python = ">=3.9"
authors = [
  {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "openai>=2.0.0",
    "anthropic",
    "pydantic>=2.0",
    "python-dotenv",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/llm-json-streaming"
"Bug Tracker" = "https://github.com/yourusername/llm-json-streaming/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-asyncio",
]
